Yoinked Siedemnastkobot's code, rebranded it a bit, and „ported” it to Bun.
As part of that porting/rebranding:
* Old .lock's had to be stashed away becasue their reliance on Canvas was breaking shit.
* The prepare script got temproarily disabled. (TSC was failing and right now - my priority was to just have it complete a `bun install`)
* A README was added
* A new Bun lockfile was created
* The .gitignore was merged from GH and Siedemnastkobot
* Changed TSConfig to use Bundler module resolution and target ES2024.
* Yanked typescript out of devDependencies in package.json becasue Bun handles TS all on its own
* Some rebranding was done to package.json (autor (which, btw, I also co-authored the original Sidemnastkobot - now, I might've been more harm than good (it was my first ever real JS project), but I did stuff nonetheless, so putting myself in the credits so soon, before this gets de-Sidemnaskoboted and re-Poltergeisted, isn't misleading, I think), version, name and description - but, importantly, NOT the license: it was originally also under MIT, so everything we're doing is perfectly legal, if a bit ethically dubious)

Everything else you see in this commit is 1:1 the same as Siedemnastkobot (or, more specifically, Siedemnastkobot's commit right before SimonB50 swooped in and fixed a lot of the mess that I caused in the codebase - becasue I felt like basing this of off that would be too unfair to SimonB50). I don't even know if this runs yet, but probably not, since (I imagine) a proper Bun port would take a lot more than that.